+++ lib/Bootstrap/Abstract.php
@@ -1,0 +1,19 @@
+<?
+class Core_Library_Bootstrap_Abstract {
+	
+	function preDetectApp(){}
+	function detectApp(){}
+	function postDetectApp(){}
+	
+	function preRoute(){}
+	function route(){}
+	function postRoute(){}
+	
+	function preAction(){}
+	function action(){}
+	function postAction(){}
+	
+	function preRender(){}
+	function render(){}
+	function postRender(){}
+}